Move: 상품문의 관련 기능들 board 패키지 내로 이동
