+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/board/api/request/PostItemInquiryRequest.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.toyproject.team4.core.item.api.request
+package com.wafflestudio.toyproject.team4.core.board.api.request
 
 import com.wafflestudio.toyproject.team4.common.CustomHttp400
 import com.wafflestudio.toyproject.team4.core.board.database.InquiryEntity
