+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/item/api/ItemController.kt
@@ -1,20 +1,12 @@
 package com.wafflestudio.toyproject.team4.core.item.api
 
-import com.wafflestudio.toyproject.team4.common.Authenticated
 import com.wafflestudio.toyproject.team4.common.CustomHttp400
-import com.wafflestudio.toyproject.team4.common.UserContext
-import com.wafflestudio.toyproject.team4.core.item.api.request.PostItemInquiryRequest
 import com.wafflestudio.toyproject.team4.core.item.api.response.ItemRankingResponse
 import com.wafflestudio.toyproject.team4.core.item.service.ItemService
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RequestHeader
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
@@ -50,18 +42,6 @@ class ItemController(
         @RequestParam index: Long?,
         @RequestParam count: Long?
     ) = itemService.getItemInquiries(itemId, index ?: 0L, count ?: 5L)
-
-    @Authenticated
-    @PostMapping("/item/{id}/inquiry")
-    fun postItemInquiry(
-        @RequestHeader(value = "Authorization") authorization: String,
-        @UserContext username: String,
-        @PathVariable(value = "id") itemId: Long,
-        @RequestBody postItemInquiryRequest: PostItemInquiryRequest
-    ) = ResponseEntity(
-        itemService.postItemInquiry(username, itemId, postItemInquiryRequest),
-        HttpStatus.CREATED
-    )
 
     @GetMapping("/search")
     fun searchItemByQuery(
