+++ src/main/kotlin/com/wafflestudio/toyproject/team4/core/item/database/ItemEntity.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.toyproject.team4.core.item.database
 
 import com.wafflestudio.toyproject.team4.core.board.database.InquiryEntity
-import com.wafflestudio.toyproject.team4.core.item.api.request.PostItemInquiryRequest
+import com.wafflestudio.toyproject.team4.core.board.api.request.PostItemInquiryRequest
 import com.wafflestudio.toyproject.team4.core.item.domain.Item
 import com.wafflestudio.toyproject.team4.core.user.database.UserEntity
 import javax.persistence.CascadeType
